+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-29 16:22:22
+
+### Changed Files
+- `src/features/hierarchy/components/AgentTable.tsx`
+
+### Statistics
+```
+ 1 file changed, 9 insertions(+), 20 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-29 16:05:10
 
 ### Changed Files